+++ package.json
@@ -12,7 +12,7 @@
     "qs": "^6.11.1",
     "vue": "^3.2.47",
     "vue-router": "^4.1.6",
-    "vuex": "^4.1.0"
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@popperjs/core": "^2.11.6",
