+++ src/App.vue
@@ -1,11 +1,21 @@
 <script setup>
-import Login from './views/Login.vue'
+import { computed } from 'vue';
+import SuccessMessage from './views/modal/SuccessMessage.vue'
+import store from './utils/vuex/store'
+
+const code = computed(() => { store.state.isLogin })
+const c = computed(() => { code.value })
+
+const message = computed(() => { store.state.successMessage })
+const m = computed(()=>{message.value})
+
 </script>
 
 <template>
   <!-- haeder -->
   <!-- content -->
   <div class="">
+    <SuccessMessage v-model="message" />
     <RouterView />
   </div>
   <!-- footer -->
