--- src/components/util/store.js
@@ -1,16 +1,0 @@
-const state = {
-    isLogin: false,
-    successMessage: ''
-};
-
-const mutations = {
-    login(state) {
-        state.isLogin = true;
-    },
-    logout(state) {
-        state.isLogin = false;
-    },
-    setSuccessMessage(state, message) {
-        state.successMessage = message;
-    }
-};
